# Conflicts:
#	docs/refs.bib

+++ skactiveml/pool/_epistemic_uncertainty_sampling.py
@@ -39,7 +39,7 @@ class EpistemicUncertaintySampling(SingleAnnotatorPoolQueryStrategy):
     missing_label : scalar or string or np.nan or None, optional
     (default=MISSING_LABEL)
         Value to represent a missing label.
-    random_state : numeric or np.random.RandomState
+    random_state : int or np.random.RandomState
         The random state to use.
 
     References
